+++ build.gradle
@@ -6,7 +6,7 @@ plugins{
 }
 
 group = "com.github.jk1"
-version = "0.3"
+version = "0.3.1"
 
 sourceCompatibility = 1.6
 targetCompatibility = 1.6
